refactor(api): turn api call into a service
